model created for hadith

+++ src/main/java/com/roze/hikmah/config/ApplicationConfig.java
@@ -45,7 +45,7 @@ public class ApplicationConfig {
         return authenticationConfiguration.getAuthenticationManager();
     }
     @Bean
-    public AuditorAware<Integer> auditorAware() {
+    public AuditorAware<Long> auditorAware() {
         return new ApplicationAuditAware();
     }
     @Bean
